added admin verification for operations

+++ DoFest/DoFest.Business/Activities/Services/Implementations/ActivityTypesService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using CSharpFunctionalExtensions;
@@ -8,6 +9,9 @@ using DoFest.Business.Activities.Services.Interfaces;
 using DoFest.Business.Errors;
 using DoFest.Entities.Activities;
 using DoFest.Persistence.Activities.ActivityTypes;
+using DoFest.Persistence.Authentication;
+using DoFest.Persistence.Authentication.Type;
+using Microsoft.AspNetCore.Http;
 
 namespace DoFest.Business.Activities.Services.Implementations
 {
@@ -15,11 +19,18 @@ namespace DoFest.Business.Activities.Services.Implementations
     {
         private readonly IMapper _mapper;
         private readonly IActivityTypesRepository _activityTypesRepository;
+        private readonly IHttpContextAccessor _accessor;
+        private readonly IUserRepository _userRepository;
+        private readonly IUserTypeRepository _userTypeRepository;
 
-        public ActivityTypesService(IMapper mapper, IActivityTypesRepository activityTypesRepository)
+        public ActivityTypesService(IMapper mapper, IActivityTypesRepository activityTypesRepository,
+            IUserRepository userRepository, IHttpContextAccessor accessor, IUserTypeRepository userTypeRepository)
         {
             _mapper = mapper;
             _activityTypesRepository = activityTypesRepository;
+            _userRepository = userRepository;
+            _accessor = accessor;
+            _userTypeRepository = userTypeRepository;
         }
 
         public async Task<Result<IEnumerable<ActivityTypeModel>, Error>> Get()
@@ -38,6 +49,14 @@ namespace DoFest.Business.Activities.Services.Implementations
                 return Result.Failure<ActivityTypeModel, Error>(ErrorsList.ActivityTypeExists);
             }
 
+            var userId = Guid.Parse(_accessor.HttpContext.User.Claims.First(c => c.Type == "userId").Value);
+            var user = await _userRepository.GetById(userId);
+            var userType = await _userTypeRepository.GetByName("Admin");
+            if (user.UserTypeId != userType.Id)
+            {
+                return Result.Failure<ActivityTypeModel, Error>(ErrorsList.UnauthorizedUser);
+            }
+
             var activityType = _mapper.Map<ActivityType>(model);
 
             await _activityTypesRepository.Add(activityType);
@@ -52,6 +71,14 @@ namespace DoFest.Business.Activities.Services.Implementations
             if (activityType == null)
             {
                 return Result.Failure<string, Error>(ErrorsList.UnavailableActivityType);
+            }
+
+            var userId = Guid.Parse(_accessor.HttpContext.User.Claims.First(c => c.Type == "userId").Value);
+            var user = await _userRepository.GetById(userId);
+            var userType = await _userTypeRepository.GetByName("Admin");
+            if (user.UserTypeId != userType.Id)
+            {
+                return Result.Failure<string, Error>(ErrorsList.UnauthorizedUser);
             }
 
             _activityTypesRepository.Delete(activityType);
